refactor: RegisterViewController 병합
