+++ app/src/test/java/com/techcourse/controller/RegisterControllerTest.java
@@ -2,6 +2,7 @@ package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import webmvc.org.springframework.web.servlet.ModelAndView;
@@ -33,5 +34,25 @@ class RegisterControllerTest {
         assertThat(modelAndView)
                 .usingRecursiveComparison()
                 .isEqualTo(new ModelAndView(new JspView("/index.jsp")));
+    }
+
+    @Test
+    @DisplayName("가입 뷰를 보여준다.")
+    void registerViewController() {
+        //given
+        final HttpServletRequest request = mock(HttpServletRequest.class);
+        final HttpServletResponse response = mock(HttpServletResponse.class);
+
+        final HttpSession session = mock(HttpSession.class);
+        when(request.getSession()).thenReturn(session);
+        when(session.getAttribute("user")).thenReturn(null);
+
+        //when
+        final ModelAndView modelAndView = registerController.execute(request, response);
+
+        //then
+        assertThat(modelAndView)
+                .usingRecursiveComparison()
+                .isEqualTo(new ModelAndView(new JspView("/register.jsp")));
     }
 }
